cmds.json fixed

+++ benchmark/insertionsort/frst.c
@@ -2,8 +2,8 @@
 #include <stdio.h>
 #include <values.h>
 
-typedef float TARGET_TYPE;
-typedef unsigned long TARGET_INDEX;
+typedef int8_t TARGET_TYPE;
+typedef uint8_t TARGET_INDEX;
 
 void insertionsort(TARGET_INDEX size, TARGET_TYPE a[size])
 {
